requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
anthropic==0.7.0
pinecone-client==2.2.4
newspaper3k==0.2.8
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.2
pydantic-settings==2.1.0
spacy==3.7.2